+++ InvenTree/web/src/src/App.tsx
@@ -15,11 +15,11 @@ import { de, en, hu } from "make-plural/plurals";
 import { useEffect, useState } from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { QrCodeModal } from './components/modals/QrCodeModal';
-import { useApiState } from './contex/ApiState';
-import { AuthProvider } from './contex/AuthContext';
-import { useLocalState } from './contex/LocalState';
-import { useSessionState } from './contex/SessionState';
-import { UserProps } from './contex/states';
+import { useApiState } from './context/ApiState';
+import { AuthProvider } from './context/AuthContext';
+import { useLocalState } from './context/LocalState';
+import { useSessionState } from './context/SessionState';
+import { UserProps } from './context/states';
 import { defaultHostList } from './defaults';
 import ErrorPage from './pages/ErrorPage';
 import { Dashboard } from './pages/Index/Dashboard';
