+++ InvenTree/web/src/src/components/nav/Header.tsx
@@ -10,8 +10,8 @@ import {
 import { useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 import { languages } from '../../App';
-import { useApiState } from '../../contex/ApiState';
-import { useLocalState } from '../../contex/LocalState';
+import { useApiState } from '../../context/ApiState';
+import { useLocalState } from '../../context/LocalState';
 import { tabs } from '../../defaults';
 import { InvenTreeStyle } from '../../globalStyle';
 import { ColorToggle } from '../items/ColorToggle';
