+++ InvenTree/web/src/src/pages/Logout.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../contex/AuthContext';
+import { useAuth } from '../context/AuthContext';
 
 export function Logout() {
   const { handleLogout } = useAuth();
